Add condition on reset of LineReferentialSyncs/StopAreaReferentialSyncs by sidekiq initializer
Refs #4314

+++ config/initializers/sidekiq.rb
@@ -1,8 +1,10 @@
 Sidekiq.configure_server do |config|
-  [
-    LineReferentialSync.pending,
-    StopAreaReferentialSync.pending
-  ].each do |pendings|
-    pendings.map { |sync| sync.failed({error: 'Failed by Sidekiq reboot', processing_time: 0}) }
+  if ENV["CHOUETTE_SIDEKIQ_CANCEL_SYNCS_ON_BOOT"]
+    [
+      LineReferentialSync.pending,
+      StopAreaReferentialSync.pending
+    ].each do |pendings|
+      pendings.map { |sync| sync.failed({error: 'Failed by Sidekiq reboot', processing_time: 0}) }
+    end
   end
 end
